fix: 사파리 버그 수정 등 - 사파리에서 텍스트 복붙 안 되던 버그, 거래 내역에 호버하면 그라디언트 안 보이던 버그 수정 - 그룹 거래 내역 빈 데이터 페이지에서 사이드바에 팁 섹션 오버레이 되던 버그 수정 - 그룹 거래 내역 응답 객체에 작성자 프로필 사진 url 필드 추가 - 새로운 버전 페이지들의 이전 버전 페이지 제거 - 토스트 ui 스택 방식으로 수정
