+++ src/app/transaction/group/page.tsx
@@ -1,5 +1,5 @@
-import GroupTransactionEmptyPageClientV2 from "@/components/transaction/group/GroupTransactionEmptyPageClientV2";
+import GroupTransactionEmptyPageClient from "@/components/transaction/group/GroupTransactionEmptyPageClient";
 
 export default function GroupTransactionEmptyPage() {
-  return <GroupTransactionEmptyPageClientV2 />;
+  return <GroupTransactionEmptyPageClient />;
 }
