+++ src/app/transaction/my/page.tsx
@@ -1,7 +1,7 @@
 // app/transaction/my/page.tsx
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
-import MyTransactionPageClientV2 from "@/components/transaction/my/MyTransactionPageClientV2";
+import MyTransactionPageClient from "@/components/transaction/my/MyTransactionPageClient";
 
 export const metadata = {
   title: "개인 거래 내역",
@@ -18,5 +18,5 @@ export default async function MyTransactionPage() {
     redirect("/");
   }
 
-  return <MyTransactionPageClientV2 />;
+  return <MyTransactionPageClient />;
 }
